Specify sensor cal. values & soil probe status via menu as settings
Turns out the convenience of swapping devices in and out without updating
serial numbers is more important than the convenience of not having to
program sensor-specific values in the field. Also, the soil moisture
probes are sufficiently unreliable that it's worth being able to remotely
disable them, so they too become settings (variables).

**Reprogram all settings incl. sensor cal. values at next deployment!**